worker/requirements.txt
livekit>=1.0.0
livekit-agents>=1.0.0
livekit-plugins-openai>=0.7.0
python-dotenv>=1.0.0
